move logic process to service layer(modify event first, if ok, i will modify post to service also)
