+++ api/v1/event.py
@@ -37,16 +37,18 @@ def create_event(event: EventCreate, db: Session = Depends(get_db)):
 @router.delete("/{event_id}")
 def delete(event_id: int, db: Session = Depends(get_db)):
     event = service.event.get_event(db, event_id)
+    if event is None:
+        raise HTTPException(status_code=404, detail="Event not found")
     return service.event.delete(db=db, event=event)
 
 
 @router.patch("/{event_id}")
 def update_event(event_id: int, event_update: EventUpdate, db: Session = Depends(get_db)):
-    existing_event = service.event.get_event(db, event_id)
+    existing_event = service.event.get_event_by_id(db, event_id)
     if existing_event is None:
         raise HTTPException(status_code=404, detail="Event not found")
-    updated_post = service.event.patch_event(db, event_id, event_update)
-    return updated_post
+    updated_event = service.event.patch(db, event_id, event_update)
+    return updated_event
 
 
 @router.post("/{event_id}/subscribers")
